Remove some event types and replace with callbacks
Since modules should depend on event queue, not vice versa, the modules
itself needs to contain the handlers, and pass those as callbacks to
the event queue.

This means the event type switch in the event library gets lot smaller,
and when the rest of the events is rewritten, it can be removed
altogheter, menaing the event library can stay on top of the other
libraries in the dependency chain.

In practice, when registering an event, pass a pointer to the event
handler method instead of a hardcoded event id. That method will be
called when time comes, and the event library doesn't even have to
bother what type of event it is.

Change-Id: If6cdef556f48d124899e5ec55074f1331a3095b7
Signed-off-by: Max Sikstrom <msikstrom@op5.com>