Remove t-tap/test_events.c - this is old
This can't really been executed since nagios 3, since it uses really
really old prototypes of the reschedule_event-method (using head and
tail indicating the old linked list structure). The event queue is up
for rewriting to use monotonic time, and this file is just messing with
autocompletion in the IDE with all the stub methods.

Removing this is no loss but some gain.

Change-Id: Ie8834603ae68553765d0d1d899d6f19074f3321b
Signed-off-by: Max Sikstrom <msikstrom@op5.com>