Refactor service check scheduling
Always reschedule a service check one check_interval in the future
directly at event handler execution. If we want something closer,
replace the event, so we can guarantee that one service check event is
scheduled at all time. If we can do that, we know for sure that we
don't forget about service checks.

Change-Id: I007313f6948784a0e77a5b5ff5e8dcadb6881760
Signed-off-by: Max Sikstrom <msikstrom@op5.com>