Rewrite event execution and creation
This guarantees that the event callback is executed once and only once.
If the event is aborted, a abort flag is set, if it is executed
normally, a normal flag is set.

That means, the callback should always clean up its user_data, but only
execute the event if normal is set.

Change-Id: I7c3e509fd5824d46317210ef672be5c5d4afca92
Signed-off-by: Max Sikstrom <msikstrom@op5.com>