Rewrite event queue to use monotonic time
The event queue shouldn't use wall time. It's nasty when administrators
starts updating the system time. To convert to monotonic time, the heap
is also simplified, with the unnessecary pqueue/squeue interface
removed, and packed with the event handling.

the pqueue was an implementation made too generic, and squeue was a
event-specific layer which stored to less information with the event.
This heap is quite small, and embedded with the event queue.

However. It would be nice to get iobroker embedded with the event
module too, so events.c can be replaced with for example libev,
libevent or possibly glib instead.

Change-Id: I9b184c85d64c88a90c837edb86788ecc58ffa4dc
Signed-off-by: Max Sikstrom <msikstrom@op5.com>