nerd: Add channels for exporting host and service performance data
These channels are vital for getting rid of one of the most
performance-hungry parts of a monitoring system, which is the I/O
load created by spooling performance data to files.

By streaming it instead, we'll be handling things far more
efficiently.

The channel exports performance data in the kvvec format, with
NULL-separated entries consisting of newline-separated fields
of key=value format, where perfdata is given as a C string.

Signed-off-by: Andreas Ericsson <ae@op5.se>