lib: Replace iocache with bufferqueue
Actually, it's a rewrite and a rename in one: the reason it's renamed
is that it's neither particularly related to IO, or a cache, and the
API is slightly different, so while not, while I'm breaking things?

A bufferqueue is simply a queue (linked list) of buffers (blocks of
memory). It does the same thing as the old iocache, but behaves in a
less cryptic way.

The iocache was "almost" a circular buffer. However, it moved all the
memory in the buffer to the start of the buffer a lot - including for
some calls that ought to be idempotent. So in theory, this made number
of memory copies O(n) for n = number of bytes in the buffer.

So, I could have gone for an actual circular buffer, to get the
"obvious" O(1) performance, but I think it fits poorly with the nature
of this structure, which is fundamentally unbounded - a read gives an
arbitrary amount of data, a send ships off an arbitrary amount of
data, and sometimes we generate too fast or read too slow, and
shouldn't start deleting data. Sure, circular buffers can be resized
(which the old iocache could), but this bufferqueue doesn't need to
be, so why go there?

Instead, the data to each nm_bq_push (or nm_bq_read) is turned into
its own buffer object, and these buffers are inserted into a regular
queue. This means each block must be copied once to get it out of the
queue - otherwise, the whole dequeued block wouldn't necessarily be
continuous in memory. To get data into the queue, nm_bq_push copies
the data as well, for memory safety reasons. However, both read and
write operations on sockets have shortcuts that remove one of the
copies each.  This leaves us at up to 2 copies - but usually 1 - for
each byte, i.e. O(1).

This is mostly irrelevant to performance, though - mostly, the buffers
are pretty empty. This is also why the O(n*m) delimiter hunt is still
in place, despite Knuth-Morris-Pratt being O(n): avoiding copies here
isn't critical.

It's more relevant that the old implementation was funny-looking,
behaved oddly, and had a hard-to-use, dangerous API.

For one thing, it could get full. This one auto-grows and auto-shrinks.

Additionally, iocache_use_delim returned a pointer to the internal
buffer, and thus the caller could (and, of course, did) twiddle with
the delimiter freely, despite being told that the delimiter wasn't
part of the returned data - an obvious offender is:

 while ((buf = iocache_use_delim(command_worker.ioc, "\n", 1, &size))) {
     buf[size] = 0;

The API also contained funky methods like unuse, which sometimes undid
the previous use, but since a lot of calls would move the next unread
data on top of where the old read data was, there was a high chance of
unuse not working unless you read the source code really, really
carefully - this implementation has peek if you need to look at the
data before committing to it.

This API tries to be more opaque, use more "standard" terminology
that should make it easier to work with, and protects its internal
data from outside view to increase memory safety. Some aspects are
trickier internally (delimiter hunt across block boundaries really
isn't a favorite), but as long as the caller doesn't have to care,
we're good.

Change-Id: Ia4e0da17fb17628a6e7680705ee9baa60905a009
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>