xpddefault: Remove use of stdio functions
Instead, go down to the basic, "unbuffered" IO layer - they're both
pretty much unbuffered, and when we call fflush all the time anyway,
it's not like we use the buffered bit of the buffered IO.

The reason for this is that the rest of naemon - the "lib" code in
particular - uses plain write for writing, rather than a higher-level
fputs. This will let us reuse code.

To make this change succinctly, remove one file opener, and two file
closer. close() is good enough for closing files, and the opening is
near-identical already.

Change-Id: Ieed5ef70208491673d3ceab43edf8f85f27beead
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>