build: Use sources rather than objects when building tests
Because we override all the build flags all the time, this does mean
that the tests will rebuild the test objects a ton of times, leading
to longer build times - we could fix this by building an archive once,
and using that instead of listing all the sources all the time, but
the tests aren't *that* expensive to build.

The reason I think this is a decent trade-off is that I can now do,
for instance, `make test-iobroker && ./test-iobroker', and if I change
the iobroker code, the iobroker code will be rebuilt - before this,
automake didn't understand that test-iobroker depended on libnaemon.la
(and thus iobroker.c), and therefore made iterative test improvements
extremely frustrating.

Change-Id: I9751d095230651366dc47aa82721f68219ec83fe
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>