tests.mk: Remove spurious -Wno-error
At some point, building the tests gave a ton of warnings. That sucked,
but was a lot of pointless work to fix, so we started building tests
with -Wno-error.

Since then, apparently someone got fed up with the warnings and fixed
almost all of them all, which means building with -Werror is now
feasible. However, the -Wno-error was left in place.

Then, during the conversion to non-recursive make, this line was
carried along, meaning that tests.mk always overrode the main
project's -Werror with a -Wno-error. This is not what anyone expected.

Change-Id: Iff747f487d3b050201d6711730cd02fa316e26db
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>