logging: Move away code that depends on objects
In my dream world, logging.c is the kind of file that everything use,
next to or just above nm_alloc - small, simple, and reusable, possibly
even for use by modules for their own files, or by workers so they'd
be able to log something about why communication with the master
failed - currently, the worker wlog method only sends a message for
the master to log.

In this world, logging depends on most parts of the system.

This at least removes all of the object dependencies from the logging
code, which is a pretty big step - it means that the rest is mostly
"platform" stuff, i.e. utils and globals, so if someone gets around to
organizing them, we should be free of deps.

Change-Id: I9f18013c6b5edfef9c841631454f4a27068be177
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>