objects_command: Separate creation and registration
This would make it possible to create an object that isn't monitored,
fiddling with it for a bit, and then (maybe) hook it up into the
global state. For instance, xodtemplate could skip about 5 of its
different phases of copying and reformatting and transformation, by
just creating the final object in the first place.

This creates some sense of symmetry between the _create and the
_destroy - destroy undoes what create creates, but it does not
unregister - in fact, there is no unregister function.

Change-Id: Ia65db0c04bc9d6150949fb8cc2e9554155e59196
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>