hashtable: Replace dkhash with GHashTable
This gives us a well defined, mature API as well as features you'd
expect from a proper hash table (like dynamic viscosity, generic
iteration, etc.). It also has the benefit of using a well studied hash
function for strings (djb), which while not super-optimised, should be
good enough for all our current (and future, I'm guessing) uses.

Of course, if it turns out to be detrimental, we can always switch it
out, because this API supports that!

Change-Id: I14c3192a2cfd41854a7127e9c9fde34c3692fe46
Signed-off-by: Anton Lofgren <alofgren@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>