configuration: Rewrite cycle detection
The code contained obscuring macros (it handled NULL pointers, but we
only really need to check for NULL pointers once, and then it's just
simple, readable lookups), and each node in the graph checked its
children rather than itself (except for dependencies, which half-way
checked their hosts?), which confused things further.

This hence tries to unwrap the cycle detection and simplify the checks
we do. In particular, this will make this structure more amendable to
being traversed as a callback, which would help us to replace the data
structure that members, etc are stored as.

This does nothing about the fact that xodtemplate has cycle detection
code as well, that it uses for the same thing but on other
properties. Oh well, another day, perhaps.

Change-Id: Ice24e06898c140b1bcdbff119bda9ee6b24a3a2a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>