objects: Clean up unused error detection code
We used to set the return code of dkhash_create in this error
variable, but I recently moved that code to its own method. Before
that, we probably added the result of malloc to this variable, but
nm_malloc never fails. Hence, the constructor should always either
create the object or log an error.

Change-Id: I70eb9bd42179e814a7af444dad0ee514eccf02ee
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>