lib: Move service_key functions to libnaemon
Since Naemon modules that are also GLib users are likely to want to
create hash tables with service keys, this is useful to expose through
the library.

I'm anticipating that we'll want to add more object (contact, host,
service, etc.) related utilities in the future so I'm adding this as a
new unit.

We really should come up with a forward plan for the library before 1.0
hits.

Change-Id: I0eef6459707dc8840806132a9c0181d3f2c78f33
Signed-off-by: Anton Lofgren <alofgren@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>