objects_service: Use more objects, less strings
Back in 93244f, I claimed that "if the caller has an object, asking it
to retrieve the name string just so [the callee] can lookup the object
we had to begin with is inefficient. If the caller doesn't have an
object, it's easy to lookup. Hence, make the function accept the
object instead of the string." This argument is just as valid here.

Change-Id: I41526bc89bf8481cfd083862c6a16c8e9105a59a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>