xrddefault: Start making crazy assumptions always
Such that, water is wet, the sun will rise tomorrow, and that a state
history that resulted in flapping before the retention data file was
written will result in the exact same thing after reading it again.

Now, I'm being glib, here - it's possible we would rewrite the
flapping algorithm (or the user changes the thresholds), so this
assumption wouldn't be true. However, the easiest to compensate for
the old flapping detection algorithm result no longer being applicable
is to do nothing at all - the next time the check is run, the flapping
algorithm will re-run, with the correct parameters, and the correct
result will be generated and the correct action will be taken.

This behavior was implemented before if you had
large_installation_tweaks - but there's no way I'm testing both those
code paths, so let's stick to the obviously correct one.

Change-Id: I82ecad0d33894914208b1f0ed10e280ce4bb79ef
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>