objects: Make constructors simple
The 30-40 parameters constructors are simply unusable - let
xodtemplate play with that if it wants to, but I'm not brave enough to
use those.

Because this is simply a constructor, the newly-created checks to make
sure that no objects that already exists are created are removed - if
I, as a programmer, want to create two hosts called "linux-server1", I
should be allowed to - though I shouldn't be allowed to register them
both into the global state.

As part of this, the services register their own relationships to the
hosts - no need to bother hosts about that. This means that the
circular dependency between hosts and services is *almost* gone -
hosts still have a pesky servicesmember member, but if someone
replaces that with a tree, the dependencies would no longer loop.

Change-Id: Ief79331201d55f39420016970c7bf3c917149ac7
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>