objects: Break object structs a bit more
Remove a few reserved variables that were kept as padding, as well as
some variables that were never, ever used.

Sure, this only saves 2-3 megabytes of memory, even on relatively
large systems, but there are *a lot* of variables in the object
structs, and having to by a 4k monitor only to see the host struct,
when the variables aren't even used, sounds silly.

Change-Id: I8870471ff06ef87e32fab6ead5abc5995492e10a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>