workers: Fix off-by-one that rendered workers useless
We would check if we were back at the last worker that received a
check at the beginning of the loop. If there is only one worker (which
is the case in merlin's mon test dist suite for me), we would simply
not run any checks at all, because the last worker is the same as the
first worker.

In addition to the single-worker special case, this also highlights a
general off-by-one: if the only available worker is the same worker
that ran the last check, that check should be allowed to run.

Change-Id: I7d4bb80b3dc57c74a2bf00d4337909a37a9c9fcb
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>