kvvec: Remove auto-length-discover behavior from methods with explicit length
This forces the user to choose if they want to use auto-length
calculation or not. Before, the kvvec would insist on "helping out"
whenever the user wanted to add 0 bytes of data from a pointer that
wasn't NULL, by reading some random memory and adding those bytes to
the vector instead of the nothingness it was told to. Now, this
behavior is optional - addkv_str if you want something simple,
addkv_wlen if you want to be explicit, and kvvec will do what you tell it.

Change-Id: Id098bde16b01c778431c412098b0100d64ada96e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>