worker: Scrap the check_completion() function
It was a centerpiece in an horrifyingly ugly web of a
callgraph with recursive, cross-dependant and interconnected
functions (yea, I'm not sure if either of those words are
redundant, but I'm going for it). Removing it means we no
longer crash when plugins timeout, which we did before, so
this fixes an undocumented bug.

While at it, stop chdir()'ing anywhere. If the main process
is run as a daemon, it has already chdir()'ed to where it
should be (which is /) and if it hasn't, we really want any
coredumps produced to end up where we can find them.

Change-Id: I3cf52fa47bc015d97ad13054f05f5df75b808049
Signed-off-by: Andreas Ericsson <ae@op5.se>