perfdata: Create the perfdata file if it doesn't exist
When we try to open(2) a file that doesn't exist and we really
want it to, we need to pass O_CREAT (Oh, cravat!), or we'll end
up complaining that the file we want to write to can't be
opened.

It was done for pipes, but not for files. This patch adds it
for the files, so with this patch we should once again produce
perfdata graphs.

Change-Id: If716a25873fd1d4599e5e53d0e149351b1c8757d
Signed-off-by: Andreas Ericsson <ae@op5.se>