downtime: Never allow adding downtime to nonexisting objects
Doing so will cause problems for loaded modules that expect
us to not be insane.

While at it, we make downtime behave like comments. That is,
if we load a downtime from retention data for an object that
has expired, we'll send a DOWNTIME_STOP event to all our
modules. That will make life a bit easier for modules that
track status outside of the naemon core (such as a report
module, where downtime actually matters).

Change-Id: Iae665a047ed07355ec0687c28c61f3c70307bf2d
Signed-off-by: Andreas Ericsson <ae@op5.se>