objects_service: Move the check command argument to the setup phase
For mocking purposes, it's important that the constructor doesn't do
any lookups into the global structures.

This makes the check command argument funky - it's required to the
setup call, but if you decide not to call setup, you don't have to set
it. I don't know what to do about it right now, though.

Change-Id: I73c2b437fbe12ba4b69aea0742fcbbd8ade33e5f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>