Fortify object lookup functions
All of naemon and a lot of its modules are riddled with
assumptions that the object lookup functions are safe to
use with NULL pointers. They were once upon a time, but
the shift to ghash made them less so.

This patch restores the fortification in the lookup
functions so that find_<object>(NULL) now once again
returns NULL.

Change-Id: I8c187045f5e3c54f54b0190fee97fd815188ca89
Signed-off-by: Andreas Ericsson <ae@op5.se>