naemon: Obsolete naemon_user & naemon_group options
These options are not very useful, since using them with a value
other than the ./configure'd one breaks a lot of stuff - permissions,
mostly. Thus, it's likely a better idea to not let users shoot
themselves in the foot with this particular gun.

This patch also adds a "sed" for the sample-config/conf.d/contacts.cfg
file which was previously missing, setting the email address to
@naemon_user@ for the sample contact.

Change-Id: I94948804a36e0121185739c073a24a3b710dc2bc
Signed-off-by: Anton Lofgren <alofgren@op5.com>