Log contact NSR's with the same format as notifications
Since some programs (showlog, import, looking at you), pretend that the
log format is a well-defined parseable format, not adhering to it causes
unjustified incompatibilites. In any case, there's precedence for
logging contact names, so let's just use that and focus on something
less depressing.

This is a prerequisite for #9686.

Change-Id: I0558fc898b4415e89d283fcc0aa132c41d510801
Signed-off-by: Anton Lofgren <alofgren@op5.com>