perfdata: Remove xpddefault
perfdata.c contained quite literally 8 lines of actual logic, plus a
bunch of thin, thin wrappers around xpddefault - not that xpddefault
didn't already expose its contents, mind you, it's just that you had
the option to choose to not use the wrapper.

This moves the wrapped functions into their wrappers, and moves the
rest of xpddefault into perfdata - it's just a straight-up copy, so
I'm not even renaming the static methods or variables, but as they're
static, they're free to rename whenever.

Change-Id: Ia7307a55585fcefd9a749cff832a6367fde0025d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>