Give some love to the contrib directory
By "some love", I actually mean "the axe". The README was wildly
outdated, and the nagios-worker.c file is obsoleted by the move
of the worker code to its own file. It's a full implementation of
an actual worker, that we actually keep up to date, so it's 100%
guaranteed to be a useful example.

Change-Id: I54f378ae1775916f40f94b05dc97a4296acfbff0
Signed-off-by: Andreas Ericsson <ae@op5.se>