Isolate the worker code to src/worker/worker.c
There's no good reason to having the entry point in one file and all
the logic in another. This patch makes src/worker/worker.c contain
everything a worker needs, including the setup, so it can be used as
an example when writing special-purpose workers.

As a bonus, it also becomes trivial to write tests that verify that
workers actually work, without having to do the setup-steps yourself.

While at it, clear out some deprecated (and dead) functions and
declare unexported functions static.

Change-Id: I6340bd9d50df662846b5fc19add71dfe2e2f43ca
Signed-off-by: Andreas Ericsson <ae@op5.se>