janitor: Remove retarded comments
If the name of the function is "get_raw_command_line()", you can
reasonably expect to "get" a/the "raw command line" from it. Any
developer who doesn't understand that *without* a comment saying
"get the raw command line" will probably not understand it *with*
the comment either.

Those of us who are not parsnips tend to find them either in the
way, annoying, insulting or any combination thereof.

Change-Id: I5c7273520b2278f8c07dc46b26b0c58e5ad59545
Signed-off-by: Andreas Ericsson <ae@op5.se>