Let environment set test output verbosity
Setting the verbosity level of checks in the test itself makes it hard
to debug test cases. When developing tests, verbose output is nice, but
almost everywhere else, normal (only faild tests) is nice to get output
from.

So just use enviornment to define test verbosity. If not defined in
enviorment, CK_NORMAL is the default, thus no change if not explicitly
asking for it during test-execution.

Change-Id: I4258f10c93ec6c46be04d0859c70fe00883d4454
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>