Make sure lockfile is unlocked before forking
This'll allow us to communicate this particular dameonization failure
back to the invoker through exit codes. This is useful for init scripts,
for example. Yes, a race exists where another process might lock between
our check and our post-fork locking attempt, but this is more of a user
experience feature than it is actual resource acquisition validation.

This finishes up MON-8039.

Change-Id: Ia926df3040dcea3153dd7143e76568f4795c5d5a
Signed-off-by: Anton Lofgren <alofgren@op5.com>