daemonizing: Remove a severely strange construct
"(pid = (pid_t)pidno) == getpid()" looks really confusing at a
quick glance, and especially when you see it in a patch. During
a recent review that didn't even modify that code I had to go
look up all the surrounding context to make sure nothing broke
simply because of this weird line.

It means "assign pidno to pid and compare the result of the
assignment to the return value of getpid()", so the correctness
of the code relies on the fact that the result of a C assignment
is the value of the assignee after the assignment is done. A lot
of things would break if that was ever changed so the code isn't
exactly wrong, but it's confusing enough that we'd better get
rid of it. To quote Anton, who summed it up nicely:
"I had a moment of great 'Huh!?' when I saw it".

Change-Id: Ica29acf2d0905609864db6168536dd9b2709d304
Signed-off-by: Andreas Ericsson <ae@op5.se>