utils: Verify that we can write PID to lockfile
In some cases (most notably, when the disk containing the lockfile is
full), we won't be able to write our PID to the lockfile. This is a
problem on systems that run supervisors that continously monitor the
lockfile in order to restart Naemon, since it will eventually lead to a
whole lot of Naemons cropping up, which is not helpful at all.

This fixes MON-8117.

Change-Id: I5104b2674c60488ecc0c85eb575c7330079f41b3
Signed-off-by: Anton Lofgren <alofgren@op5.com>