test_macros: Refactor object creation somewhat
This patch just makes it easier to see the real changes, that come
in this series by disentangling the object variables and putting
them one-by-one on their own lines.

As part of the refactoring, we make the object structures static,
meaning variables that aren't set are always initialized to NULL,
in order to avoid surprises in any code that handles it, and we
make a TEST_HOSTNAME macro that we use instead of printing in the
same string over and over. This will be used more in the next
patch, where the test first failed because I had misspelled the
hostname.

Change-Id: Id2107929cc49dcfb48cae7d8794b26834a9432fd
Signed-off-by: Andreas Ericsson <ae@op5.se>