test_macros: Add tests for ondemand macros
Originally, this test triggered the "crash on NULL in object
lookups" bug that was fixed in the previous commit, because
the error was first discovered through misuse of ondemand
macros.

However, the tests are still valid and now that they pass
we should keep them.

This tests the "steps to reproduce" for MON-8127.

Change-Id: I2b3eaa81ee23a6f3eff6a451d2f2ef5c3fc2c532
Signed-off-by: Andreas Ericsson <ae@op5.se>