host checks: Up the current_attempt when results are processed
That's how services do things, and it's how we have to handle
things in order to make sure two peered nodes behave the same
way.

This *should* fix MON-7053, but in order to write tests for it,
I would need to refactor the check result processing code which
I'm reluctant to do so close to a release.

One drawback of this patch is that the $HOSTCHECKATTEMPT$ macro
will now yield "we've tried this many times so far" instead of
"this is the Xth attempt", but that's consistent with how it's
done for services, so I'm not sure that's a drawback.

Change-Id: I990dc1c89346a17e6103d2d3af12652ab7a5c0b4
Signed-off-by: Andreas Ericsson <ae@op5.se>