Start refactoring check result processing
The previous code was a complete mess of unused parameters and
setting storage mixed with actions.

This is the first step towards making sure state setting is
handled uniformly and in a verifiable manner for all host
checks. It also allows us to run some further tests to the
post-processing phase of check running as well as paves the
way for the titillating check result transformer, as the
refactoring provides a natural breakpoint between setting
all the object's variables and acting upon them.

A lot still remains though, such as handling flapping and
state history at the same time as we're setting all other
variables that a checkresult ends up affecting.

As a neat sideeffect (and the actual reason for the refactoring)
is that this resolves MON-7053.

Change-Id: Ie3b167ac167b788c774a7a9c01b5a9c5ab1dcb5a
Signed-off-by: Andreas Ericsson <ae@op5.se>