Make init script use existing sysvinit infrastructure
Reinventing this particular wheel has been done enough time by all
the different distros. Let's not do it again. This should be
compatible with redhat (and derivatives), SuSE (and derivatives),
and debian (and derivatives).

Now we actually wait for naemon to be stopped before returning
from a `service naemon stop`, or equivalent invocation.

While at it, we also make sure to return the correct value to
the user, since that's sort of important for scripts that use
this script.

Signed-off-by: Andreas Ericsson <ae@op5.se>